backend/pyproject.toml
[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
aiofiles = "^23.1.0"
python-multipart = "^0.0.6"
aioredis = "^2.0.1"
httpx = "^0.24.0"
pydantic = "^1.10.7"
PyPDF2 = "^3.0.0"
python-docx = "^0.8.11"
pandas = "^2.1.0"
sqlalchemy = "^2.1.0"            # optional, for DB
alembic = "^1.12.0"              # optional, for migrations
google-generativeai = "^0.2.0"   # Gemini Flash 2.5 API SDK
python-dotenv = "^1.0.1"         # load .env
